Restructure widget files and update the title
