+++ src/lib/module/home_settings/widget/settings_body/widget/swipe_to_app.dart
@@ -1,4 +1,4 @@
-part of 'settings_body.dart';
+part of '../settings_body.dart';
 
 class _SwipeToApp
     extends ThemedGetWidget<UserApplicationsController, HomeTheme> {
