+++ src/lib/module/home_settings/widget/settings_body/widget/vibration_preference.dart
@@ -1,4 +1,4 @@
-part of 'settings_body.dart';
+part of '../settings_body.dart';
 
 class _VibrationPreferences
     extends ThemedGetWidget<UserApplicationsController, HomeTheme> {
